backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pyyaml==6.0.1
requests==2.32.0
python-multipart==0.0.9
pandas==2.2.0
